backend/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
tensorflow==2.19.0
Pillow==9.5.0
numpy>=1.26,<2.2